requirements.txt
# Slack integration
slack_bolt>=1.18.0
slack_sdk>=3.27.0
aider-chat
pydantic
strands-agents>=1.10.0
strands-agents-tools>=0.2.9
sympy
typing
pandas
numpy
requests
beautifulsoup4
lxml
sentencepiece
scikit-learn
nltk
spacy
transformers>=4.44.0
tensorboardX>=2.5
torch
scipy>=1.13.0
setuptools>=75.0.0
rank_bm25
tavily-python>=0.3.0
readability-lxml>=0.8.1


# Testing dependencies
pytest>=7.0.0
pytest-asyncio>=0.21.0
pytest-mock>=3.10.0
pytest-cov>=4.0.0
pytest-benchmark>=4.0.0
pytest-timeout>=2.1.0
